pyproject.toml
[tool.poetry]
name = "chat-streamlit-practice"
version = "0.1.0"
description = ""
authors = ["Jiho Ko"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3.9"
langchain-core = "^0.3.21"
langchain-community = "^0.3.9"
langchain-experimental = "^0.3.3"
langchain-openai = "^0.2.10"
streamlit = "^1.40.2"
python-dotenv = "^1.0.1"
duckduckgo-search = "^6.3.7"
wikipedia = "^1.4.0"


[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
